move helper function to best_sections
