+++ wizer/tests/integration_tests/test_import_demo_activities.py
@@ -1,5 +1,5 @@
 from wizer import models
-from wizer.file_importer import _activity_suitable_for_awards
+from wizer.best_sections.fastest import _activity_suitable_for_awards
 
 
 def test_import_of_demo_activities(import_demo_data, client):
@@ -108,3 +108,35 @@ def test_import_of_activities__evaluates_for_awards(import_one_activity, tracks_
 
     # some best sections got saved
     assert models.BestSection.objects.filter(activity=activity).count() > 0
+
+
+def test__activity_evaluates_for_awards(insert_activity):
+    # in case both activity and sport are suitable, function should also return True
+    activity = insert_activity(name="dummy-1", evaluates_for_awards=True)
+    assert activity.evaluates_for_awards is True
+    assert activity.sport.evaluates_for_awards is True
+    assert _activity_suitable_for_awards(activity=activity) is True
+
+    # in case only activity is not suitable, function should return False
+    activity = insert_activity(name="dummy-2", evaluates_for_awards=False)
+    assert activity.evaluates_for_awards is False
+    assert activity.sport.evaluates_for_awards is True
+    assert _activity_suitable_for_awards(activity=activity) is False
+
+    # in case only sport is not suitable, function should return False
+    activity = insert_activity(name="dummy", evaluates_for_awards=True)
+    assert activity.evaluates_for_awards is True
+    sport = activity.sport
+    sport.evaluates_for_awards = False
+    sport.save()
+    assert activity.sport.evaluates_for_awards is False
+    assert _activity_suitable_for_awards(activity=activity) is False
+
+    # in case both sport and activity are not suitable, function should also return False
+    activity = insert_activity(name="dummy", evaluates_for_awards=False)
+    assert activity.evaluates_for_awards is False
+    sport = activity.sport
+    sport.evaluates_for_awards = False
+    sport.save()
+    assert activity.sport.evaluates_for_awards is False
+    assert _activity_suitable_for_awards(activity=activity) is False
